# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "CareTogether CMS API",
    "description": "API for the CareTogether Community Management System",
    "contact": {
      "name": "CareTogether CMS Team",
      "url": "https://caretogether.io/"
    },
    "license": {
      "name": "Use under AGPLv3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/{organizationId}/{locationId}/Files/{documentId}": {
      "get": {
        "tags": [
          "Files"
        ],
        "operationId": "Files_GetReadValetUrl",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "operationId": "Files_GenerateUploadValetUrl",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/{organizationId}/{locationId}/Referrals": {
      "get": {
        "tags": [
          "Referrals"
        ],
        "operationId": "Referrals_ListAllReferrals",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Referral"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Referrals/referralCommand": {
      "post": {
        "tags": [
          "Referrals"
        ],
        "operationId": "Referrals_SubmitReferralCommand",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralCommand"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Referrals/arrangementCommand": {
      "post": {
        "tags": [
          "Referrals"
        ],
        "operationId": "Referrals_SubmitArrangementCommand",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArrangementCommand"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referral"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Person": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Age"
              }
            ]
          }
        }
      },
...
      "ContactInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "personId": {
            "type": "string",
            "format": "guid"
          },
          "addresses": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          },
          "currentAddressId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "phoneNumbers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PhoneNumber"
            }
          },
          "preferredPhoneNumberId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "emailAddresses": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/EmailAddress"
            }
          },
          "preferredEmailAddressId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "contactMethodPreferenceNotes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Address": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "line1": {
            "type": "string",
            "nullable": true
          },
          "line2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "stateId": {
            "type": "string",
            "format": "guid"
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "PhoneNumber": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/PhoneNumberType"
          }
        }
      },
...
      "EmailAddress": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/EmailAddressType"
          }
        }
      },
...
      "Referral": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "policyVersion": {
            "type": "string",
            "nullable": true
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "closeReason": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ReferralCloseReason"
              }
            ]
          },
          "partneringFamily": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Family"
              }
            ]
          },
          "contacts": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ContactInfo"
            }
          },
          "referralFormUploads": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FormUploadInfo"
            }
          },
          "referralActivitiesPerformed": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ActivityInfo"
            }
          },
          "arrangements": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Arrangement"
            }
          }
        }
      },
      "ReferralCloseReason": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "NotAppropriate",
          "Resourced",
          "NoCapacity",
          "NoLongerNeeded",
          "NeedMet"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4
        ]
      },
      "Family": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "volunteerFamilyStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/VolunteerFamilyStatus"
              }
            ]
          },
          "partneringFamilyStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PartneringFamilyStatus"
              }
            ]
          },
          "adults": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ValueTupleOfPersonAndFamilyAdultRelationshipInfo"
            }
          },
          "children": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "custodialRelationships": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustodialRelationship"
            }
          }
        }
      },
      "VolunteerFamilyStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Active",
          "Inactive"
        ],
        "enum": [
          0,
          1
        ]
      },
      "PartneringFamilyStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Active",
          "Inactive"
        ],
        "enum": [
          0,
          1
        ]
      },
      "ValueTupleOfPersonAndFamilyAdultRelationshipInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "item1": {
            "$ref": "#/components/schemas/Person"
          },
          "item2": {
            "$ref": "#/components/schemas/FamilyAdultRelationshipInfo"
          }
        }
      },
...
      "CustodialRelationship": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "childId": {
            "type": "string",
            "format": "guid"
          },
          "personId": {
            "type": "string",
            "format": "guid"
          },
          "type": {
            "$ref": "#/components/schemas/CustodialRelationshipType"
          }
        }
      },
...
      "FormUploadInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string",
            "format": "guid"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "formName": {
            "type": "string",
            "nullable": true
          },
          "formVersion": {
            "type": "string",
            "nullable": true
          },
          "originalFileName": {
            "type": "string",
            "nullable": true
          },
          "uploadedDocumentId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "ActivityInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string",
            "format": "guid"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "activityName": {
            "type": "string",
            "nullable": true
          },
          "performedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "performedByPersonId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "Arrangement": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "policyVersion": {
            "type": "string",
            "nullable": true
          },
          "arrangementType": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/ArrangementState"
          },
          "arrangementFormUploads": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FormUploadInfo"
            }
          },
          "arrangementActivitiesPerformed": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ActivityInfo"
            }
          },
          "volunteerAssignments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/VolunteerAssignment"
            }
          },
          "partneringFamilyChildAssignments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PartneringFamilyChildAssignment"
            }
          },
          "childrenLocationHistory": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ChildrenLocationHistoryEntry"
            }
          },
          "notes": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Note"
            }
          }
        }
      },
      "ArrangementState": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Setup",
          "Open",
          "Closed"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "VolunteerAssignment": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "FamilyVolunteerAssignment": "#/components/schemas/FamilyVolunteerAssignment",
            "IndividualVolunteerAssignment": "#/components/schemas/IndividualVolunteerAssignment"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "arrangementFunction": {
            "type": "string",
            "nullable": true
          },
          "discriminator": {
            "type": "string"
          }
        }
      },
...
      "PartneringFamilyChildAssignment": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "personId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "ChildrenLocationHistoryEntry": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string",
            "format": "guid"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "childrenIds": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string",
              "format": "guid"
            }
          },
          "familyId": {
            "type": "string",
            "format": "guid"
          },
          "plan": {
            "$ref": "#/components/schemas/ChildrenLocationPlan"
          },
          "additionalExplanation": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ReferralCommand": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "CloseReferral": "#/components/schemas/CloseReferral",
            "CreateReferral": "#/components/schemas/CreateReferral",
            "PerformReferralActivity": "#/components/schemas/PerformReferralActivity",
            "UploadReferralForm": "#/components/schemas/UploadReferralForm"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "referralId": {
            "type": "string",
            "format": "guid"
          },
          "discriminator": {
            "type": "string"
          }
        }
      },
      "CloseReferral": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReferralCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "closeReason": {
                "$ref": "#/components/schemas/ReferralCloseReason"
              }
            }
          }
        ]
      },
      "CreateReferral": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReferralCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "familyId": {
                "type": "string",
                "format": "guid"
              },
              "policyVersion": {
                "type": "string",
                "nullable": true
              },
              "openedAtUtc": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "PerformReferralActivity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReferralCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "activityName": {
                "type": "string",
                "nullable": true
              },
              "performedAtUtc": {
                "type": "string",
                "format": "date-time"
              },
              "performedByPersonId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "UploadReferralForm": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReferralCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "formName": {
                "type": "string",
                "nullable": true
              },
              "formVersion": {
                "type": "string",
                "nullable": true
              },
              "uploadedFileName": {
                "type": "string",
                "nullable": true
              },
              "uploadedDocumentId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "ArrangementCommand": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "AssignIndividualVolunteer": "#/components/schemas/AssignIndividualVolunteer",
            "AssignPartneringFamilyChildren": "#/components/schemas/AssignPartneringFamilyChildren",
            "AssignVolunteerFamily": "#/components/schemas/AssignVolunteerFamily",
            "CreateArrangement": "#/components/schemas/CreateArrangement",
            "EndArrangement": "#/components/schemas/EndArrangement",
            "InitiateArrangement": "#/components/schemas/InitiateArrangement",
            "PerformArrangementActivity": "#/components/schemas/PerformArrangementActivity",
            "TrackChildrenLocationChange": "#/components/schemas/TrackChildrenLocationChange",
            "UploadArrangementForm": "#/components/schemas/UploadArrangementForm"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "referralId": {
            "type": "string",
            "format": "guid"
          },
          "arrangementId": {
            "type": "string",
            "format": "guid"
          },
          "discriminator": {
            "type": "string"
          }
        }
      },
      "AssignIndividualVolunteer": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "personId": {
                "type": "string",
                "format": "guid"
              },
              "arrangementFunction": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "AssignPartneringFamilyChildren": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "childrenIds": {
                "type": "array",
                "nullable": true,
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        ]
      },
      "AssignVolunteerFamily": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "familyId": {
                "type": "string",
                "format": "guid"
              },
              "arrangementFunction": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "CreateArrangement": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "policyVersion": {
                "type": "string",
                "nullable": true
              },
              "arrangementType": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "EndArrangement": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "endedAtUtc": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "InitiateArrangement": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "initiatedAtUtc": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "PerformArrangementActivity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "activityName": {
                "type": "string",
                "nullable": true
              },
              "performedAtUtc": {
                "type": "string",
                "format": "date-time"
              },
              "performedByPersonId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "TrackChildrenLocationChange": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "changedAtUtc": {
                "type": "string",
                "format": "date-time"
              },
              "childrenIds": {
                "type": "array",
                "nullable": true,
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              },
              "familyId": {
                "type": "string",
                "format": "guid"
              },
              "plan": {
                "$ref": "#/components/schemas/ChildrenLocationPlan"
              },
              "additionalExplanation": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "UploadArrangementForm": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ArrangementCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "formName": {
                "type": "string",
                "nullable": true
              },
              "formVersion": {
                "type": "string",
                "nullable": true
              },
              "uploadedFileName": {
                "type": "string",
                "nullable": true
              },
              "uploadedDocumentId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
...
    }
  }
}